+++ src/layout/BsMenu.vue
@@ -1,0 +1,12 @@
+<script setup>
+  import BsButton from '../components/BsButton.vue';
+  import { faBars } from '@fortawesome/free-solid-svg-icons'
+  import FontAwesomeIcon from '../components/FontAwesomeIcon.vue';
+
+</script>
+
+<template>
+  <div>
+    <FontAwesomeIcon style="font-size: 2rem;" :icon="faBars" />
+  </div>
+</template>