+++ src/layout/Header.vue
@@ -1,12 +1,19 @@
 <script setup>
-  import DarkMode from './DarkMode.vue';
+  import { inject } from 'vue'
+  import DarkMode from '../components/DarkMode.vue'
+  import BsMenu from './BsMenu.vue'
   import { RouterLink } from 'vue-router'
+
+  const appSettings = inject('appSettings')
 </script>
 
 <template>
-  <header>
+  <header :style="{maxWidth: appSettings.layoutWidth.value}">
     <RouterLink class="logo" to="portfolio">BS</RouterLink>
-    <DarkMode />
+    <div style="display: flex; gap: 10px; align-items: center">
+      <DarkMode />
+      <BsMenu />
+    </div>
   </header>
 </template>
 
@@ -15,6 +22,9 @@
     display: flex;
     align-items: center;
     justify-content: space-between;
+    margin-inline: auto;
+    width: 100%;
+    transition: 200ms ease-out;
   }
 
   header .logo {