+++ src/pages/Home.vue
@@ -1,15 +1,13 @@
 <template>
-  <div class="content-wrapper">
-    <h1>Hello, I'm Brett</h1>
-    <p>I'm a Senior Frontend Web Developer, and this is where I plan to host some personal projects.</p>
-    <h4>Coming Soon</h4>
-    <ol>
-      <li><span style="text-decoration: line-through">Dark Mode Toggle</span></li>
-      <li><router-link to="hypno-eye">Hypno Eye 👁</router-link></li>
-      <li>SSFP (Searchable, Sortable, Filterable, Paginated) Datatable</li>
-      <li>D3 Interactive Charts &amp; Maps</li>
-      <li>Various CSS Artboards...</li>
-      <li>and much more!</li>
-    </ol>
-  </div>
+  <h1>Hello, I'm Brett</h1>
+  <p>I'm a Senior Frontend Web Developer, and this is where I plan to host some personal projects.</p>
+  <h4>Coming Soon</h4>
+  <ol>
+    <li><router-link to="hypno-eye">Hypno Eye 👁</router-link></li>
+    <li><router-link to="color-comparison">Color Comparison 🎨</router-link></li>
+    <li>SSFP (Searchable, Sortable, Filterable, Paginated) Datatable</li>
+    <li>D3 Interactive Charts &amp; Maps</li>
+    <li>Various CSS Artboards...</li>
+    <li>and much more!</li>
+  </ol>
 </template>