+++ src/routes.js
@@ -1,6 +1,7 @@
 import { createRouter, createWebHashHistory } from 'vue-router'
 import Home from './pages/Home.vue'
 import HypnoEye from './pages/HypnoEye.vue'
+import ColorComparison from './pages/ColorComparison.vue'
 
 const routes = [
   {
@@ -14,6 +15,10 @@ const routes = [
   {
     path: '/hypno-eye',
     component: HypnoEye
+  },
+  {
+    path: '/color-comparison',
+    component: ColorComparison
   },
 ]